Add list, insert, update, delete features with client/server validation and error handling

+++ src/main/webapp/list.jsp
@@ -1,3 +1,5 @@
+<%@ page import="java.util.*,com.myproject.crudapp.model.entity.Employee"%>
+
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -22,18 +24,54 @@
             </tr>
         </thead>
         <tbody>
+
+           <%
+           String msg1 = request.getParameter("msg1");
+           if(msg1!=null){
+           %>
+           <p style="color:red"><%= msg1%></p>
+           <%}%>
+
+          <%
+           String msg2 = request.getParameter("msg2");
+           if(msg2!=null){
+           %>
+           <p style="color:green"><%= msg2%></p>
+           <%}%>
+
+           <%
+           String msg3 = request.getParameter("msg3");
+           if(msg3!=null){
+           %>
+           <p style="color:green"><%= msg3%></p>
+           <%}%>
+
+            <%
+            int cnt = 1;
+            List<Employee> elist = (List<Employee>)request.getAttribute("list");
+            if(elist!=null && !elist.isEmpty()){
+              for(Employee emp : elist){
+            %>
             <tr>
-                <td>1</td>
-                <td>Om Ghodekar</td>
-                <td>om@gmail.com</td>
-                <td>it</td>
-                <td>fullstack developer</td>
-                <td>28000</td>
+                <td><%= cnt++%></td>
+                <td><%= emp.getName()%></td>
+                <td><%= emp.getEmail()%></td>
+                <td><%= emp.getDepartment()%></td>
+                <td><%= emp.getDesignation()%></td>
+                <td><%= emp.getSalary()%></td>
                 <td>
-                    <a href="employee?action=edit">update</a>
-                    <a href="employee?action=delete">delete</a>
+                    <a href="employee?action=edit&id=<%= emp.getId()%>">update</a>
+                    <a href="employee?action=delete&id=<%= emp.getId()%>" onclick="return confirm('delete record') ">delete</a>
                 </td>
             </tr>
+            <%}
+            }else{%>
+            <tr>
+               <td colspan="7">
+                  <p>no records found</p>
+               </td>
+            </tr>
+            <%}%>
         </tbody>
     </table>
 </body>
